+++ checkitbaby/fortigate_agent.py
@@ -96,7 +96,7 @@ class Fortigate_agent(Agent, fortigate_agent_ipsec.Mixin, fortigate_agent_execut
         elif data['group'] == 'route':
             result = self.process_route(line=line, agent=data['agent'], conn=data['conn'], type=data['type'], check=data['check'], command=data['command'])
         elif data['group'] == 'sdwan':
-            result = self.process_sdwan(line)
+            result = self.process_sdwan(line=line, agent=data['agent'], conn=data['conn'], type=data['type'], check=data['check'], command=data['command'])
         else:
             log.error("Syntax error: unknown command group {}".format(data['command']))
             raise SystemExit
@@ -187,156 +187,6 @@ class Fortigate_agent(Agent, fortigate_agent_ipsec.Mixin, fortigate_agent_execut
                 else:
                    log.error("Could not connect to FortiGate {} but continue scenario ".format(self.name))
 
-    def _check_bgp_requirement(self, result={}, rname='', rvalue=''):
-        """
-        Validates bgp routes requirements, that is verifying if the 'has ...' part
-        """
-        log.info("Enter with rname={} rvalue={} result={}".format(rname, rvalue, result))
-        fb = True  # by default, requirement is met
-        if not 'subnet' in result:
-            result['subnet'] = []
-        if not 'nexthop' in result:
-            result['nexthop'] = []
-        if not 'interface' in result:
-            result['interface'] = []
-        if rname in ('total', 'recursive', 'subnet','nexthop','interface'):
-            log.debug("requirement {}={} is known".format(rname, rvalue))
-            if rname == 'total' :
-                log.debug("Checking exact number of subnets : asked={} got={}".format(rvalue, result['total']))
-                if str(result['total']) == str(rvalue):
-                    log.debug("Total number of bgp routes is matching requirement")
-                else:
-                    log.debug("Total number of bgp routes does not match requirement")
-                    fb = False
-            elif rname == 'recursive' :
-                log.debug("Checking exact number of recursive routes specifically : asked={} got={}".format(rvalue, result['recursive']))
-                if str(result['recursive']) == str(rvalue):
-                    log.debug("Total number of bgp recursive routes is matching requirement")
-                else:
-                    log.debug("Total number of bgp recursive routes does not match requirement")
-                    fb = False
-            elif result[rname].count(rvalue) > 0:
-                log.debug("rname={} found : requirement is met".format(rname))
-                fb = True
-            else :
-                log.debug("rname={} found : requirement is not met".format(rname))
-                fb = False
-        else:
-           log.error("unknown bgp requirement {}={} is unknown".format(rname, rvalue))
-           raise SystemExit
-        log.debug("requirements verdict : {}".format(fb))
-        return fb
-
-    def _cmd_check_sdwan(self, name="", line=""):
-        """
-        Checks on sdwan feature
-        - member selected from sdwan rule :
-
-            FGT-B1-1 # diagnose sys virtual-wan-link service 1
-			Service(1): Address Mode(IPV4) flags=0x0
-			  Gen(1), TOS(0x0/0x0), Protocol(0: 1->65535), Mode(sla)
-			  Service role: standalone
-			  Member sub interface:
-			  Members:
-				1: Seq_num(1 vpn_isp1), alive, sla(0x1), cfg_order(0), cost(0), selected
-				2: Seq_num(2 vpn_isp2), alive, sla(0x1), cfg_order(1), cost(0), selected
-				3: Seq_num(3 vpn_mpls), alive, sla(0x1), cfg_order(2), cost(0), selected
-			  Src address:
-					10.0.1.0-10.0.1.255
-
-			  Dst address:
-					10.0.2.0-10.0.2.255
-
-			FGT-B1-1 #
-
-            - check alive members :
-                ex : FGT-B1-1:1 check [sdwan_1_member1_alive] sdwan service 1 member 1 has status=alive
-            - check sla value for a particular member
-                ex : FGT-B1-1:1 check [sdwan_1_member1_sla] sdwan service 1 member 1 has sla=0x1
-            - check preferred member
-                ex : FGT-B1-1:1 check [sdwan_1_preferred] sdwan service 1 member 1 has preferred=1
-            - vdom supported :
-                ex : F1B2:1 check [sdwan_1_member1_alive] sdwan vdom=customer service 1 member 1 has status=alive
-	    """
-        log.info("Enter with name={} line={}".format(name, line))
-        found_flag = False
-        # vdom processing if needed
-        line = self._vdom_processing(line=line)
-        version = '6.4'
-        if re.search('version=6.2', line):
-            line = re.sub("version=6.2\s", '', line)
-            version = '6.2'
-        # remove requirements from line (after 'has ...')
-        line_no_requirement = line
-        line_no_requirement = line.split('has')[0]
-        log.debug("line_no_requirement={}".format(line_no_requirement))
-        match_command = re.search("sdwan\s+service\s(?P<rule>\d+)\s+member\s+(?P<seq>\d+)", line_no_requirement)
-        if match_command:
-            rule = match_command.group('rule')
-            seq = match_command.group('seq')
-            log.debug("matched sdwan service rule={} member seq={}".format(rule, seq))
-            self.connect_if_needed()
-            # Query SDWAN service
-
-            result = self._ssh.get_sdwan_service(service=rule, version=version)
-            if result:
-                if len(result['members']) >= 1:
-                    log.debug("At least 1 member was found")
-                    found_flag = True
-            # Without any further requirements, result is pass
-            feedback = found_flag
-            # Processing further requirements (has ...)
-            match_has = re.search("\s+has\s+(?P<requirements>.+)",line)
-            if match_has:
-                requirements = match_has.group('requirements')
-                log.debug("requirements list: {}".format(requirements))
-                rnum = 0
-                for r in requirements.split():
-                    rnum = rnum + 1
-                    log.debug("requirement num={} : {}".format(rnum, r))
-                    match_req = re.search("^(?P<rname>.+)=(?P<rvalue>.+)", r)
-                    if match_req:
-                        rname = match_req.group('rname')
-                        rvalue = match_req.group('rvalue')
-                        log.debug("Checking requirement {}={}".format(rname, rvalue))
-                        rfdb = self._check_sdwan_service_requirement(seq=seq, rname=rname, rvalue=rvalue, result=result)
-                        feedback = feedback and rfdb
-            self.add_report_entry(check=name, result=feedback)
-            return found_flag
-
-    def _check_sdwan_service_requirement(self, seq='', result={}, rname='', rvalue=''):
-        """
-        Validates all requirements for sdwan service
-        """
-        log.info("Enter with seq={} rname={} rvalue={} result={}".format(seq, rname, rvalue, result))
-        # if command failed, no members could be found => fail
-        if result['members'] == {}:
-            log.warning("could not retrieve sdwan members, return Fail")
-            return False
-        fb = True  # by default, requirement is met
-        # Checking member is preferred
-        # Extract seq from 1st member and see if its our
-        if rname == 'preferred':
-            if '1' in result['members']:
-                pref_member_seq = result['members']['1']['seq_num']
-                log.debug("Preferred member seq_num={}".format(pref_member_seq))
-                if pref_member_seq == seq:
-                    log.debug("requirement is fullfilled".format(rvalue))
-                    fb = True
-                else:
-                    log.debug("requirement is not fullfilled")
-                    fb = False
-            else:
-                log.error("No members found in 1st position")
-                fb = False
-        elif rname in ('status','sla'):
-            log.debug("Accepted requirement rname={} checking member={} detail={}".format(rname, seq, result['members'][seq]))
-            fb = self.check_requirement(result=result['members'][seq], name=rname, value=rvalue)
-        else:
-            log.error("unknown requirement")
-            raise SystemExit
-        log.debug("requirements verdict : {}".format(fb))
-        return fb
 
 
 if __name__ == '__main__': #pragma: no cover
